feat: reorganizing application
If we think about it - it's not very large application and it can be written like monolyth.
