--- cmd/datasource/utils.go
@@ -1,47 +1,0 @@
-package main
-
-import (
-	"log"
-	"math"
-	"time"
-
-	"github.com/gin-gonic/gin"
-	amqp "github.com/rabbitmq/amqp091-go"
-)
-
-func connectToRabbitMQ(connectionString string) (*amqp.Connection, error) {
-	var counts int64
-	var backOff = 1 * time.Second
-	var connection *amqp.Connection
-
-	for {
-		c, err := amqp.Dial(connectionString)
-		if err != nil {
-			log.Println("RabbitMQ not yet ready...")
-			counts++
-		} else {
-			log.Println("Connected to RabbitMQ")
-			connection = c
-			break
-		}
-
-		if counts > 5 {
-			log.Println(err)
-			return nil, err
-		}
-
-		backOff = time.Duration(math.Pow(float64(counts), 2)) * time.Second
-		log.Println("backing off...")
-		time.Sleep(backOff)
-		continue
-	}
-
-	return connection, nil
-}
-
-func ErrorResponse(c *gin.Context, status int, error_text string) {
-	c.JSON(status, gin.H{
-		"error":   true,
-		"message": error_text,
-	})
-}
