--- cmd/notifications/api_handlers.go
@@ -1,15 +1,0 @@
-package main
-
-import "github.com/gin-gonic/gin"
-
-func (app *App) GetNotifications(c *gin.Context) {
-
-}
-
-func (app *App) PatchNotifications(c *gin.Context) {
-
-}
-
-func (app *App) DeleteNotifications(c *gin.Context) {
-
-}
