--- cmd/router/tmp/index.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <button onclick="sendRequest()">send</button>
-    <script>
-        let socket = new WebSocket("ws://127.0.0.1:8080/ws");
-
-        socket.onopen = function (e) {
-            console.log("[open] Connection established");
-            console.log("Sending to server");
-            socket.send("My name is John");
-        };
-
-        socket.onmessage = function (event) {
-            console.log(`[message] Data received from server: ${event.data}`);
-        };
-
-        socket.onclose = function (event) {
-            if (event.wasClean) {
-                console.log(`[close] Connection closed cleanly, code=${event.code} reason=${event.reason}`);
-            } else {
-                console.log('[close] Connection died');
-            }
-        };
-
-        socket.onerror = function (error) {
-            console.log(`[error]`);
-        };
-
-        function sendRequest() {
-            socket.send("This is test message");
-        }
-
-    </script>
-</body>
-
-</html>
