+++ go.mod
@@ -1,0 +1,12 @@
+module github.com/shaninalex/financial-analyzer
+
+go 1.21.5
+
+require (
+	github.com/goccy/go-json v0.10.2
+	github.com/google/uuid v1.5.0
+	github.com/gorilla/websocket v1.5.1
+	github.com/rabbitmq/amqp091-go v1.9.0
+)
+
+require golang.org/x/net v0.17.0 // indirect
