+++ pkg/datasource/demo_data/AAPL_gurufocus_financials.json
