+++ pkg/datasource/demo_data/alphavantage_series.json
