+++ pkg/notifications/utils.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"log"
+)
+
+func failOnError(err error) {
+	if err != nil {
+		log.Panicf("%s", err)
+	}
+}
